+++ local_e2e/pkg/fixtures/manifests.go
