+++ pkg/fixture/endpointgroupbinding.go
